requirements.txt
google-api-core==1.26.1
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-api-python-client==2.0.2
google-cloud-core==1.5.0
google-cloud==0.34.0
google-cloud-bigquery==2.6.2
google-cloud-logging==2.1.1
google-cloud-pubsub==2.2.0
google-cloud-speech==2.0.1
google-cloud-storage==1.35.0
grpc-google-cloud-pubsub-v1==0.14.0
grpc-google-iam-v1==0.12.3
requests==2.25.1
pubsub==0.1.2
vosk==0.3.21
sounddevice==0.4.1
srt==3.4.1
numpy==1.20
wget==3.2
elasticsearch-dsl==7.3.0
pyaudio==0.2.11
python-dotenv==0.15.0